Fmu ref 1 (#188)
* work in progress

* work in progress

* work in progress

* work in progress

* Add after solve methods (#185)


* work in progress

* fmu terminate added after solve that contains fmus

* linters fix

* linters fix

* bb fmu test marked excluded

* bb fmu test marked excluded

* bb fmu test marked excluded

* test_bounsing_ball_2_way_pos

* test_bounsing_ball_2_way_pos

+++ numerous/engine/system/fmu_subsystem.py
@@ -274,11 +274,11 @@ class FMU_Subsystem(Subsystem, EquationBase):
 
         class EventInfo(ctypes.Structure):
 
-            _fields_ = [('newDiscreteStatesNeeded', ctypes.c_int),
-                        ('terminateSimulation', ctypes.c_int),
-                        ('nominalsOfContinuousStatesChanged', ctypes.c_int),
-                        ('valuesOfContinuousStatesChanged', ctypes.c_int),
-                        ('nextEventTimeDefined', ctypes.c_int),
+            _fields_ = [('newDiscreteStatesNeeded', ctypes.c_int8),
+                        ('terminateSimulation', ctypes.c_int8),
+                        ('nominalsOfContinuousStatesChanged', ctypes.c_int8),
+                        ('valuesOfContinuousStatesChanged', ctypes.c_int8),
+                        ('nextEventTimeDefined', ctypes.c_int8),
                         ('nextEventTime', ctypes.c_double)]
 
         event_info = EventInfo()
